separate nodes.js file into separate files #9 + add config.js with HSL colors and languages
