+++ src/configs.js
@@ -1,0 +1,15 @@
+export const languageOptions = [
+        "bg", "ca", "zh", "zh-TW", "cs", "da", "nl", "en", "en-US", "en-AU", "en-GB", "en-NZ", "en-IN",
+        "et", "fi", "nl-BE", "fr", "fr-CA", "de", "de-CH", "el", "hi", "hu", "id", "it", "ja", "ko",
+        "lv", "lt", "ms", "no", "pl", "pt", "pt-BR", "ro", "ru", "sk", "es", "sv", "th", "tr", "uk", "vi"
+    ]
+
+export const colors = {
+    indigo: "hsl(275, 100%, 25%)",
+    green: "hsl(145, 63%, 40%)",
+    greenLight: "hsl(145, 63%, 50%)",
+    red: "hsl(359, 96%, 50%)",
+    yellow: "hsl(60, 74%, 45%)",
+    yellowDark: "hsl(60, 74%, 22%)",
+    orange: "hsl(28, 80%, 52%)"
+};
