+++ src/nodes/TextToSpeechNode.js
@@ -1,0 +1,20 @@
+import { LiteGraph } from '@comfyorg/litegraph';
+import { colors } from '../configs';
+
+class TextToSpeechNode extends LiteGraph.LGraphNode {
+    constructor() {
+        super();
+        this.title = "Text to Speech";
+        this.addInput("text", "string");
+        this.addOutput("audio", "audio");
+        this.color = colors.yellowDark;
+    }
+
+    onExecute() {
+        const text = this.getInputData(0);
+        // Implement TTS logic here
+        this.setOutputData(0, { type: "audio", data: "synthesized_audio_data" });
+    }
+}
+
+export { TextToSpeechNode };
